Вынесение модели перевозчика, правка экрана с ошибками
