+++ TravelSchedule/Models/CarrierOption.swift
@@ -1,0 +1,18 @@
+//
+//  CarrierOption.swift
+//  TravelSchedule
+//
+//  Модель перевозчиков
+
+import Foundation
+
+struct CarrierOption: Identifiable, Hashable {
+    let id = UUID()
+    let carrierName: String
+    let logoName: String
+    let dateText: String
+    let depart: String
+    let arrive: String
+    let durationText: String
+    let transferNote: String?
+}
